Adding logistic regression for binary outcome and other GLM estimators (#164)
* added glm and logistic regression

* updated with tests
